+++ semcore/radio/src/Radio.jsx
@@ -3,7 +3,7 @@ import createComponent, { Component, CONTEXT_COMPONENT, sstyled, Root } from '@s
 import { Flex, Box } from '@semcore/flex-box';
 import assignProps, { callAllEventHandlers } from '@semcore/utils/lib/assignProps';
 import keyboardFocusEnhance from '@semcore/utils/lib/enhances/keyboardFocusEnhance';
-import resolveColor from '@semcore/utils/lib/color';
+import resolveColorEnhance from '@semcore/utils/lib/enhances/resolveColorEnhance';
 import getInputProps, { inputProps } from '@semcore/utils/lib/inputProps';
 
 import style from './style/radio.shadow.css';
@@ -132,7 +132,7 @@ class ValueRoot extends Component {
     includeInputProps: inputProps,
     defaultChecked: false,
   };
-  static enhance = [keyboardFocusEnhance()];
+  static enhance = [keyboardFocusEnhance(), resolveColorEnhance()];
   static displayName = 'Value';
   static contextType = RadioGroup[CONTEXT_COMPONENT];
   static style = style;
@@ -183,6 +183,7 @@ class ValueRoot extends Component {
       tag,
       disabled,
       includeInputProps,
+      resolveColor,
       ...other
     } = this.asProps;
     const [, radioMarkProps] = getInputProps(other, includeInputProps);
@@ -195,6 +196,7 @@ class ValueRoot extends Component {
       state,
       keyboardFocused,
       disabled,
+      resolveColor,
       ...propsWithoutChildren,
     };
 
@@ -248,7 +250,7 @@ Control.displayName = 'Control';
 
 const RadioMark = (props) => {
   const SValue = Root;
-  const { theme, styles } = props;
+  const { theme, styles, resolveColor } = props;
 
   return sstyled(styles)(<SValue render={Box} tag='div' use:theme={resolveColor(theme)} />);
 };