+++ website/docs/data-display/venn-chart/examples/venn-orientation.tsx
@@ -19,7 +19,7 @@ export default () => {
       <Plot height={300} width={400} data={data}>
         <Venn orientation={orientations[orientation]} orientationOrder={orders[order]}>
           <Venn.Circle dataKey='F' name='F' />
-          <Venn.Circle dataKey='S' name='S' color={colors['blue-03']} />
+          <Venn.Circle dataKey='S' name='S' />
           <Venn.Intersection dataKey='F/S' name='F/S' />
         </Venn>
       </Plot>