+++ semcore/spin-container/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [7.10.0] - 2023-10-26
+
+### Added
+
+- Design tokens resolving for `theme` prop.
+
 ## [7.9.2] - 2023-10-24
 
 ### Changed