+++ semcore/tooltip/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [6.9.0] - 2023-10-26
+
+### Added
+
+- Design tokens resolving for `theme` prop.
+
 ## [6.8.8] - 2023-10-24
 
 ### Changed